nl2sql-agent/requirements.txt
google-adk>=0.1.0
google-cloud-aiplatform>=1.43.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
vertexai>=1.0.0
pandas>=2.0.0
pyyaml>=6.0